requirements.txt
flask==2.0.1
flask-cors==3.0.10
sentence-transformers==2.2.2
google-generativeai==0.3.0
numpy==1.21.0